Changing Episodes & Assets Table Schema
So there is no need for assets to know what they are related too, I was just being dumb. Additionally there is lots of content that might not fit in the titles 45 character limit, same with the description so now the description is a text field and title is 90 characters.